Revert #135: workaround for race condition
PR #135 was a workaround for a race condition present in METplus<5 (discussed in #132).  My understanding is that spack-stack 1.9.3 has METplus>6.0, which doesn't have that race condition, so the workaround is unnecessary.

I have a vague memory of this workaround being insufficient at some point, meaning global-workflow metp jobs went back to being serial, but I can't find that discussion now.  The fix in METplus>=5 should be more general.